refactor: ProductManager에서 새상품 추가 로직 분리하여 ProductForm에 추가
